Fix version resolution - invoke getNextVersion function
The getNextVersion function stored in project.extra was not being
invoked, only converted to string which gave the function signature
instead of the actual version number.

This caused Maven artifacts to be published with incorrect names like:
com.ditchoom-buffer-fun Setup_gradle.getNextVersion...

Now properly casts and invokes the function to get the version.